WIP: Split calculations from plotting
