+++ arviz/plots/backends/matplotlib/distplot.py
@@ -1,0 +1,5 @@
+"""
+Matplotlib Backbend for distplot
+"""
+
+
