+++ arviz/plots/kdeplot.py
@@ -1,13 +1,11 @@
 """One-dimensional kernel density estimate plots."""
 import warnings
 import numpy as np
-import matplotlib.pyplot as plt
 from scipy.signal import gaussian, convolve, convolve2d  # pylint: disable=no-name-in-module
 from scipy.sparse import coo_matrix
 import xarray as xr
 from ..data import InferenceData
 from ..utils import conditional_jit, _stack
-from .plot_utils import _scale_fig_size
 from ..stats.stats_utils import histogram
 
 
@@ -148,567 +146,81 @@ def plot_kde(
         >>> az.plot_kde(mu_posterior, values2=tau_posterior, contour=False)
 
     """
+
+    if isinstance(values, xr.Dataset):
+        raise ValueError(
+            "Xarray dataset object detected.Use plot_posterior, plot_density, plot_joint"
+            "or plot_pair instead of plot_kde"
+        )
+    if isinstance(values, InferenceData):
+        raise ValueError(" Inference Data object detected. Use plot_posterior instead of plot_kde")
+
+    if values2 is None:
+        density, lower, upper = _fast_kde(values, cumulative, bw)
+
+        if cumulative:
+            density_q = density
+        else:
+            density_q = density.cumsum() / density.sum()
+
+        # This is just a hack placeholder for now
+        xmin, xmax, ymin, ymax, gridsize = [None]*5
+    else:
+        gridsize = (128, 128) if contour else (256, 256)
+        density, xmin, xmax, ymin, ymax = _fast_kde_2d(values, values2, gridsize=gridsize)
+
+        lower, upper, density_q = [None]*3
+
+
+    kde_plot_args = dict(
+        # Internal API
+        density=density,
+        lower=lower,
+        upper=upper,
+        density_q=density_q,
+        xmin=xmin,
+        xmax=xmax,
+        ymin=ymin,
+        ymax=ymax,
+        gridsize=gridsize,
+
+        # User Facing API that can be simplified
+        values=values,
+        values2=values2,
+        # cumulative=cumulative,
+        rug=rug,
+        label=label,
+        # bw=bw,
+        quantiles=quantiles,
+        rotated=rotated,
+        contour=contour,
+        fill_last=fill_last,
+        textsize=textsize,
+        plot_kwargs=plot_kwargs,
+        fill_kwargs=fill_kwargs,
+        rug_kwargs=rug_kwargs,
+        contour_kwargs=contour_kwargs,
+        contourf_kwargs=contourf_kwargs,
+        pcolormesh_kwargs=pcolormesh_kwargs,
+        ax=ax,
+        legend=legend,
+
+
+    )
     if backend is None or backend.lower() in ("mpl", "matplotlib"):
+        from .backends.matplotlib.distplot import _plot_kde_mpl
         ax = _plot_kde_mpl(
-            values=values,
-            values2=values2,
-            cumulative=cumulative,
-            rug=rug,
-            label=label,
-            bw=bw,
-            quantiles=quantiles,
-            rotated=rotated,
-            contour=contour,
-            fill_last=fill_last,
-            textsize=textsize,
-            plot_kwargs=plot_kwargs,
-            fill_kwargs=fill_kwargs,
-            rug_kwargs=rug_kwargs,
-            contour_kwargs=contour_kwargs,
-            contourf_kwargs=contourf_kwargs,
-            pcolormesh_kwargs=pcolormesh_kwargs,
-            ax=ax,
-            legend=legend,
+          **kde_plot_args
         )
     elif backend == "bokeh":
+        from .backends.bokeh.distplot import _plot_kde_bokeh
         ax = _plot_kde_bokeh(
-            values=values,
-            values2=values2,
-            cumulative=cumulative,
-            rug=rug,
-            label=label,
-            bw=bw,
-            quantiles=quantiles,
-            rotated=rotated,
-            contour=contour,
-            fill_last=fill_last,
-            textsize=textsize,
-            plot_kwargs=plot_kwargs,
-            fill_kwargs=fill_kwargs,
-            rug_kwargs=rug_kwargs,
-            contour_kwargs=contour_kwargs,
-            contourf_kwargs=contourf_kwargs,
-            pcolormesh_kwargs=pcolormesh_kwargs,
-            ax=ax,
-            legend=legend,
+            **kde_plot_args
         )
     else:
         raise NotImplementedError(
             'Backend {} not implemented. Use {{"matplotlib", "bokeh"}}'.format(backend)
         )
-    return ax
-
-    return ax
-
-
-def _plot_kde_mpl(
-    values,
-    values2=None,
-    cumulative=False,
-    rug=False,
-    label=None,
-    bw=4.5,
-    quantiles=None,
-    rotated=False,
-    contour=True,
-    fill_last=True,
-    textsize=None,
-    plot_kwargs=None,
-    fill_kwargs=None,
-    rug_kwargs=None,
-    contour_kwargs=None,
-    contourf_kwargs=None,
-    pcolormesh_kwargs=None,
-    ax=None,
-    legend=True,
-):
-    """1D or 2D KDE plot taking into account boundary conditions.
-
-    Parameters
-    ----------
-    values : array-like
-        Values to plot
-    values2 : array-like, optional
-        Values to plot. If present, a 2D KDE will be estimated
-    cumulative : bool
-        If true plot the estimated cumulative distribution function. Defaults to False.
-        Ignored for 2D KDE
-    rug : bool
-        If True adds a rugplot. Defaults to False. Ignored for 2D KDE
-    label : string
-        Text to include as part of the legend
-    bw : float
-        Bandwidth scaling factor for 1D KDE. Should be larger than 0. The higher this number the
-        smoother the KDE will be. Defaults to 4.5 which is essentially the same as the Scott's
-        rule of thumb (the default rule used by SciPy).
-    quantiles : list
-        Quantiles in ascending order used to segment the KDE. Use [.25, .5, .75] for quartiles.
-        Defaults to None.
-    rotated : bool
-        Whether to rotate the 1D KDE plot 90 degrees.
-    contour : bool
-        If True plot the 2D KDE using contours, otherwise plot a smooth 2D KDE. Defaults to True.
-    fill_last : bool
-        If True fill the last contour of the 2D KDE plot. Defaults to True.
-    textsize: float
-        Text size scaling factor for labels, titles and lines. If None it will be autoscaled based
-        on figsize.
-    plot_kwargs : dict
-        Keywords passed to the pdf line of a 1D KDE.
-    fill_kwargs : dict
-        Keywords passed to the fill under the line (use fill_kwargs={'alpha': 0} to disable fill).
-        Ignored for 2D KDE
-    rug_kwargs : dict
-        Keywords passed to the rug plot. Ignored if rug=False or for 2D KDE
-        Use `space` keyword (float) to control the position of the rugplot. The larger this number
-        the lower the rugplot.
-    contour_kwargs : dict
-        Keywords passed to ax.contour. Ignored for 1D KDE.
-    contourf_kwargs : dict
-        Keywords passed to ax.contourf. Ignored for 1D KDE.
-    pcolormesh_kwargs : dict
-        Keywords passed to ax.pcolormesh. Ignored for 1D KDE.
-    ax : matplotlib axes
-    legend : bool
-        Add legend to the figure. By default True.
-
-    Returns
-    -------
-    ax : matplotlib axes
-
-    Examples
-    --------
-    Plot default KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> import arviz as az
-        >>> non_centered = az.load_arviz_data('non_centered_eight')
-        >>> mu_posterior = np.concatenate(non_centered.posterior["mu"].values)
-        >>> az.plot_kde(mu_posterior)
-
-
-    Plot KDE with rugplot
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, rug=True)
-
-
-    Plot a cumulative distribution
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, cumulative=True)
-
-
-
-    Rotate plot 90 degrees
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, rotated=True)
-
-
-    Plot 2d contour KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> tau_posterior = np.concatenate(non_centered.posterior["tau"].values)
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior)
-
-    Remove fill for last contour in 2d KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior, fill_last=False)
-
-    Plot 2d smooth KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior, contour=False)
-
-    """
-    if ax is None:
-        ax = plt.gca()
-
-    figsize = ax.get_figure().get_size_inches()
-
-    figsize, *_, xt_labelsize, linewidth, markersize = _scale_fig_size(figsize, textsize, 1, 1)
-
-    if isinstance(values, xr.Dataset):
-        raise ValueError(
-            "Xarray dataset object detected.Use plot_posterior, plot_density, plot_joint"
-            "or plot_pair instead of plot_kde"
-        )
-    if isinstance(values, InferenceData):
-        raise ValueError(" Inference Data object detected. Use plot_posterior instead of plot_kde")
-
-    if values2 is None:
-        if plot_kwargs is None:
-            plot_kwargs = {}
-        plot_kwargs.setdefault("color", "C0")
-
-        default_color = plot_kwargs.get("color")
-
-        if fill_kwargs is None:
-            fill_kwargs = {}
-
-        fill_kwargs.setdefault("color", default_color)
-
-        if rug_kwargs is None:
-            rug_kwargs = {}
-        rug_kwargs.setdefault("marker", "_" if rotated else "|")
-        rug_kwargs.setdefault("linestyle", "None")
-        rug_kwargs.setdefault("color", default_color)
-        rug_kwargs.setdefault("space", 0.2)
-
-        plot_kwargs.setdefault("linewidth", linewidth)
-        rug_kwargs.setdefault("markersize", 2 * markersize)
-
-        density, lower, upper = _fast_kde(values, cumulative, bw)
-
-        rug_space = max(density) * rug_kwargs.pop("space")
-
-        x = np.linspace(lower, upper, len(density))
-
-        if cumulative:
-            density_q = density
-        else:
-            density_q = density.cumsum() / density.sum()
-        fill_func = ax.fill_between
-        fill_x, fill_y = x, density
-        if rotated:
-            x, density = density, x
-            fill_func = ax.fill_betweenx
-
-        ax.tick_params(labelsize=xt_labelsize)
-
-        if rotated:
-            ax.set_xlim(0, auto=True)
-            rug_x, rug_y = np.zeros_like(values) - rug_space, values
-        else:
-            ax.set_ylim(0, auto=True)
-            rug_x, rug_y = values, np.zeros_like(values) - rug_space
-
-        if rug:
-            ax.plot(rug_x, rug_y, **rug_kwargs)
-
-        if quantiles is not None:
-            fill_kwargs.setdefault("alpha", 0.75)
-
-            idx = [np.sum(density_q < quant) for quant in quantiles]
-
-            fill_func(
-                fill_x,
-                fill_y,
-                where=np.isin(fill_x, fill_x[idx], invert=True, assume_unique=True),
-                **fill_kwargs
-            )
-        else:
-            fill_kwargs.setdefault("alpha", 0)
-            if fill_kwargs.get("alpha") == 0:
-                ax.plot(x, density, label=label, **plot_kwargs)
-                fill_func(fill_x, fill_y, **fill_kwargs)
-            else:
-                ax.plot(x, density, **plot_kwargs)
-                fill_func(fill_x, fill_y, label=label, **fill_kwargs)
-        if legend and label:
-            ax.legend()
-    else:
-        if contour_kwargs is None:
-            contour_kwargs = {}
-        contour_kwargs.setdefault("colors", "0.5")
-        if contourf_kwargs is None:
-            contourf_kwargs = {}
-        if pcolormesh_kwargs is None:
-            pcolormesh_kwargs = {}
-
-        gridsize = (128, 128) if contour else (256, 256)
-
-        density, xmin, xmax, ymin, ymax = _fast_kde_2d(values, values2, gridsize=gridsize)
-        g_s = complex(gridsize[0])
-        x_x, y_y = np.mgrid[xmin:xmax:g_s, ymin:ymax:g_s]
-
-        ax.grid(False)
-        ax.set_xlim(xmin, xmax)
-        ax.set_ylim(ymin, ymax)
-        if contour:
-            qcfs = ax.contourf(x_x, y_y, density, antialiased=True, **contourf_kwargs)
-            qcs = ax.contour(x_x, y_y, density, **contour_kwargs)
-            if not fill_last:
-                qcfs.collections[0].set_alpha(0)
-                qcs.collections[0].set_alpha(0)
-        else:
-            ax.pcolormesh(x_x, y_y, density, **pcolormesh_kwargs)
-
-    return ax
-
-
-def _plot_kde_bokeh(
-    values,
-    values2=None,
-    cumulative=False,
-    rug=False,
-    label=None,
-    bw=4.5,
-    quantiles=None,
-    rotated=False,
-    contour=True,
-    fill_last=True,
-    textsize=None,
-    plot_kwargs=None,
-    fill_kwargs=None,
-    rug_kwargs=None,
-    contour_kwargs=None,
-    contourf_kwargs=None,
-    pcolormesh_kwargs=None,
-    ax=None,
-    legend=True,
-):
-    """1D or 2D KDE plot taking into account boundary conditions.
-
-    Parameters
-    ----------
-    values : array-like
-        Values to plot
-    values2 : array-like, optional
-        Values to plot. If present, a 2D KDE will be estimated
-    cumulative : bool
-        If true plot the estimated cumulative distribution function. Defaults to False.
-        Ignored for 2D KDE
-    rug : bool
-        If True adds a rugplot. Defaults to False. Ignored for 2D KDE
-    label : string
-        Text to include as part of the legend
-    bw : float
-        Bandwidth scaling factor for 1D KDE. Should be larger than 0. The higher this number the
-        smoother the KDE will be. Defaults to 4.5 which is essentially the same as the Scott's
-        rule of thumb (the default rule used by SciPy).
-    quantiles : list
-        Quantiles in ascending order used to segment the KDE. Use [.25, .5, .75] for quartiles.
-        Defaults to None.
-    rotated : bool
-        Whether to rotate the 1D KDE plot 90 degrees.
-    contour : bool
-        If True plot the 2D KDE using contours, otherwise plot a smooth 2D KDE. Defaults to True.
-    fill_last : bool
-        If True fill the last contour of the 2D KDE plot. Defaults to True.
-    textsize: float
-        Text size scaling factor for labels, titles and lines. If None it will be autoscaled based
-        on figsize.
-    plot_kwargs : dict
-        Keywords passed to the pdf line of a 1D KDE.
-    fill_kwargs : dict
-        Keywords passed to the fill under the line (use fill_kwargs={'alpha': 0} to disable fill).
-        Ignored for 2D KDE
-    rug_kwargs : dict
-        Keywords passed to the rug plot. Ignored if rug=False or for 2D KDE
-        Use `space` keyword (float) to control the position of the rugplot. The larger this number
-        the lower the rugplot.
-    contour_kwargs : dict
-        Keywords passed to ax.contour. Ignored for 1D KDE.
-    contourf_kwargs : dict
-        Keywords passed to ax.contourf. Ignored for 1D KDE.
-    pcolormesh_kwargs : dict
-        Keywords passed to ax.pcolormesh. Ignored for 1D KDE.
-    ax : matplotlib axes
-    legend : bool
-        Add legend to the figure. By default True.
-
-    Returns
-    -------
-    ax : matplotlib axes
-
-    Examples
-    --------
-    Plot default KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> import arviz as az
-        >>> non_centered = az.load_arviz_data('non_centered_eight')
-        >>> mu_posterior = np.concatenate(non_centered.posterior["mu"].values)
-        >>> az.plot_kde(mu_posterior)
-
-
-    Plot KDE with rugplot
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, rug=True)
-
-
-    Plot a cumulative distribution
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, cumulative=True)
-
-
-
-    Rotate plot 90 degrees
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, rotated=True)
-
-
-    Plot 2d contour KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> tau_posterior = np.concatenate(non_centered.posterior["tau"].values)
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior)
-
-    Remove fill for last contour in 2d KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior, fill_last=False)
-
-    Plot 2d smooth KDE
-
-    .. plot::
-        :context: close-figs
-
-        >>> az.plot_kde(mu_posterior, values2=tau_posterior, contour=False)
-
-    """
-    import bokeh.plotting as bkp
-
-    if ax is None:
-        ax = bkp.figure(sizing_mode="stretch_both")
-
-    # TODO: handle
-    # figsize = ax.get_figure().get_size_inches()
-
-    # figsize, *_, xt_labelsize, linewidth, markersize = _scale_fig_size(figsize, textsize, 1, 1)
-
-    if isinstance(values, xr.Dataset):
-        raise ValueError(
-            "Xarray dataset object detected.Use plot_posterior, plot_density, plot_joint"
-            "or plot_pair instead of plot_kde"
-        )
-    if isinstance(values, InferenceData):
-        raise ValueError(" Inference Data object detected. Use plot_posterior instead of plot_kde")
-
-    if values2 is None:
-        if plot_kwargs is None:
-            plot_kwargs = {}
-        plot_kwargs.setdefault("line_color", plt.rcParams["axes.prop_cycle"].by_key()["color"][0])
-
-        default_color = plot_kwargs.get("color")
-
-        if fill_kwargs is None:
-            fill_kwargs = {}
-
-        fill_kwargs.setdefault("color", default_color)
-
-        if rug_kwargs is None:
-            rug_kwargs = {}
-        # rug_kwargs.setdefault("marker", "_" if rotated else "|")
-        # rug_kwargs.setdefault("linestyle", "None")
-        # rug_kwargs.setdefault("color", default_color)
-        # rug_kwargs.setdefault("space", 0.2)
-
-        # plot_kwargs.setdefault("line_width", linewidth)
-        # rug_kwargs.setdefault("markersize", 2 * markersize)
-
-        density, lower, upper = _fast_kde(values, cumulative, bw)
-
-        # rug_space = max(density) * rug_kwargs.pop("space")
-
-        x = np.linspace(lower, upper, len(density))
-
-        if cumulative:
-            density_q = density
-        else:
-            density_q = density.cumsum() / density.sum()
-
-        ax.line(x, density, **plot_kwargs)
-        """
-        fill_func = ax.fill_between
-        fill_x, fill_y = x, density
-        if rotated:
-            x, density = density, x
-            fill_func = ax.fill_betweenx
-
-        ax.tick_params(labelsize=xt_labelsize)
-
-        if rotated:
-            ax.set_xlim(0, auto=True)
-            rug_x, rug_y = np.zeros_like(values) - rug_space, values
-        else:
-            ax.set_ylim(0, auto=True)
-            rug_x, rug_y = values, np.zeros_like(values) - rug_space
-
-        if rug:
-            ax.plot(rug_x, rug_y, **rug_kwargs)
-
-        if quantiles is not None:
-            fill_kwargs.setdefault("alpha", 0.75)
-
-            idx = [np.sum(density_q < quant) for quant in quantiles]
-
-            fill_func(
-                fill_x,
-                fill_y,
-                where=np.isin(fill_x, fill_x[idx], invert=True, assume_unique=True),
-                **fill_kwargs
-            )
-        else:
-            fill_kwargs.setdefault("alpha", 0)
-            if fill_kwargs.get("alpha") == 0:
-                ax.plot(x, density, label=label, **plot_kwargs)
-                fill_func(fill_x, fill_y, **fill_kwargs)
-            else:
-                ax.plot(x, density, **plot_kwargs)
-                fill_func(fill_x, fill_y, label=label, **fill_kwargs)
-        """
-    else:
-        # todo
-        raise NotImplementedError("Sorry")
-        if contour_kwargs is None:
-            contour_kwargs = {}
-        contour_kwargs.setdefault("colors", "0.5")
-        if contourf_kwargs is None:
-            contourf_kwargs = {}
-        if pcolormesh_kwargs is None:
-            pcolormesh_kwargs = {}
-
-        gridsize = (128, 128) if contour else (256, 256)
-
-        density, xmin, xmax, ymin, ymax = _fast_kde_2d(values, values2, gridsize=gridsize)
-        g_s = complex(gridsize[0])
-        x_x, y_y = np.mgrid[xmin:xmax:g_s, ymin:ymax:g_s]
-
-        ax.grid(False)
-        ax.set_xlim(xmin, xmax)
-        ax.set_ylim(ymin, ymax)
-        if contour:
-            qcfs = ax.contourf(x_x, y_y, density, antialiased=True, **contourf_kwargs)
-            qcs = ax.contour(x_x, y_y, density, **contour_kwargs)
-            if not fill_last:
-                qcfs.collections[0].set_alpha(0)
-                qcs.collections[0].set_alpha(0)
-        else:
-            ax.pcolormesh(x_x, y_y, density, **pcolormesh_kwargs)
-
     return ax
